Fix compile warnings
Reviewed-on: http://gerrit.lab.local:8080/1125
Smoketests: Mike McLaughlin <mmclaughlin@scalecomputing.com>
Tested-by: Mike McLaughlin <mmclaughlin@scalecomputing.com>

+++ unittest/FStringUnitTest.cpp
@@ -251,7 +251,7 @@ TEST_F(FStringTest, ImplodeExplodeBinaryFStringTest)
 
     ASSERT_EQ(components.size(), componentsTest.size());
 
-    for (int i=0; i < components.size(); i++)
+    for (int i=0; i < (int)components.size(); i++)
     {
         ASSERT_STREQ( components[i], componentsTest[i] );
     }
